debug模式不弹出加密狗提示

+++ src/model/default/receive/subscriptions.ts
@@ -298,22 +298,30 @@ export default {
      */
     dogWarn() {
 
-        ipcRenderer.on('dog-warn', (_: IpcRendererEvent, msg: string) => {
+        ipcRenderer.on('dog-warn', async (_: IpcRendererEvent, msg: string) => {
 
-            Modal.destroyAll();
-            Modal.confirm({
-                title: '警告',
-                content: msg ?? '后台服务异常中断或加密狗被拔出，请重启应用',
-                centered: true,
-                okText: '重新启动',
-                cancelText: '退出',
-                onOk() {
-                    ipcRenderer.send('do-relaunch');
-                },
-                onCancel() {
-                    ipcRenderer.send('do-close');
-                }
-            });
+            let isDebug = false;
+            try {
+                isDebug = await helper.isDebug();
+            } catch (error) {
+                isDebug = false;
+            }
+            if (!isDebug) {
+                Modal.destroyAll();
+                Modal.confirm({
+                    title: '警告',
+                    content: msg ?? '后台服务异常中断或加密狗被拔出，请重启应用',
+                    centered: true,
+                    okText: '重新启动',
+                    cancelText: '退出',
+                    onOk() {
+                        ipcRenderer.send('do-relaunch');
+                    },
+                    onCancel() {
+                        ipcRenderer.send('do-close');
+                    }
+                });
+            }
         });
     }
 }
